BUG: ticket #1952, allow floating ddof in std and var.
Also return nan for resulting non-positive degrees of freedom.

+++ numpy/core/src/multiarray/calculation.h
@@ -41,7 +41,7 @@ PyArray_Std(PyArrayObject* self, int axis, int rtype, PyArrayObject* out,
 
 NPY_NO_EXPORT PyObject *
 __New_PyArray_Std(PyArrayObject *self, int axis, int rtype, PyArrayObject *out,
-                  int variance, int num);
+                  int variance, double num);
 
 NPY_NO_EXPORT PyObject*
 PyArray_Sum(PyArrayObject* self, int axis, int rtype, PyArrayObject* out);
